+++ lib/takwimu/instruments/ree_gc.rb
@@ -1,5 +1,3 @@
-#!/usr/bin/env ruby
-
 # Copyright (c) 2017 Salesforce
 # Copyright (c) 2009 37signals, LLC
 #
@@ -23,15 +21,39 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
 
-require "bundler/setup"
-require "takwimu"
+module Takwimu
+  module Instruments
+    class Ruby18GC
+      def initialize
+        GC.enable_stats
+      end
 
-# You can add fixtures and/or initialization code here to make experimenting
-# with your gem easier. You can also use a different console, if you like.
+      def start!(state)
+        state[:ruby18_gc] = current
+      end
 
-# (If you use this, don't forget to add pry to your Gemfile!)
-# require "pry"
-# Pry.start
+      def instrument!(state, counters, gauges)
+        last = state[:ruby18_gc]
+        cur = state[:ruby18_gc] = current
 
-require "irb"
-IRB.start(__FILE__)
+        counters.update \
+          :'GC.count'             => cur[:gc_count] - before[:gc_count],
+          :'GC.time'              => cur[:gc_time] - before[:gc_time],
+          :'GC.memory'            => cur[:gc_memory] - before[:gc_memory],
+          :'GC.allocated_objects' => cur[:objects] - before[:objects]
+
+        gauges[:'Objects.live'] = ObjectSpace.live_objects
+        gauges[:'GC.growth'] = GC.growth
+      end
+
+      private def current
+        {
+          :objects   => ObjectSpace.allocated_objects,
+          :gc_count  => GC.collections,
+          :gc_time   => GC.time,
+          :gc_memory => GC.allocated_size
+        }
+      end
+    end
+  end
+end