+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+- Rename to takwimu 
+- Add Additional Configuration to Metrics
+- Added periodic passenger phusion stats
+
 ## 0.0.4
 
 - Support Puma backlog #14